Move DrawUnit to separate module
