+++ src/draw/mod.rs
@@ -1,13 +1,15 @@
 mod batch;
 mod pass;
 mod primitive;
+mod draw_unit;
 
 pub mod plot;
 pub mod shadow;
 
 pub use golem::Texture;
 
-pub use batch::{Batch, DrawUnit, LineBatch, TriBatch};
+pub use draw_unit::DrawUnit;
+pub use batch::{Batch, LineBatch, TriBatch};
 pub use pass::{ColPass, TexColPass};
 pub use primitive::{
     ColVertex, Geometry, GeometryMode, Line, Quad, TexColVertex, TexVertex, Triangle, Vertex,
